+++ monkeytypes/__init__.py
@@ -2,10 +2,16 @@
 Used for a common things between agent and island
 """
 from .operating_system import OperatingSystem
-from . import base_models
+
+from .base_models import (
+    InfectionMonkeyBaseModel,
+    InfectionMonkeyModelConfig,
+    MutableInfectionMonkeyBaseModel,
+    MutableInfectionMonkeyModelConfig,
+)
 from .agent_plugin_type import AgentPluginType
 from .agent_plugin_manifest import PluginName, PluginVersion, AgentPluginManifest
-from .concurrency import Lock, Event
+from .concurrency import BasicLock, Event, Lock, RLock
 from .serialization import JSONSerializable
 from .ids import AgentID, HardwareID, MachineID
 from .int_range import IntRange