+++ CHANGELOG.md
@@ -7,9 +7,23 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
 
-## [Unreleased]
+## [v0.2.0]
 ### Added
+- Installation and usage to the README
+- Support for type checking
+- Export BasicLock
+- Export RLock
+- Export InfectionMonkeyBaseModel
+- Export MutableInfectionMonkeyBaseModel
+- Export InfectionMonkeyModelConfig
+- Export MutableInfectionMonkeyModelConfig
+
 ### Changed
+- The way base models are exported
+
 ### Fixed
+
 ### Removed
+- `monkeytypes.base_models` export
+
 ### Security